Lapindrome -DSA learning series
Lapindrome is defined as a string which when split in the middle, gives two halves having the same characters and same frequency of each character.
worst case complexity can be  o(10^5). therefore, used the concept of dividing string into half ,sorting it and then comparing character by character in both halves.